VideoRec: Made uHz 32-bit.
svn:sync-xref-src-repo-rev: r112987

+++ src/VBox/Main/src-client/DrvAudioVideoRec.cpp
@@ -1,4 +1,4 @@
-/* $Id: DrvAudioVideoRec.cpp 65389 2017-01-20 14:03:51Z andreas.loeffler@oracle.com $ */
+/* $Id: DrvAudioVideoRec.cpp 65391 2017-01-20 15:29:23Z andreas.loeffler@oracle.com $ */
 /** @file
  * Video recording audio backend for Main.
  */
@@ -76,7 +76,7 @@ typedef struct AVRECCODEC
             /** Encoder we're going to use. */
             OpusEncoder    *pEnc;
             /** The encoding rate to use. */
-            uint16_t        uHz;
+            uint32_t        uHz;
             /** Duration of the frame in samples (per channel).
              *  Valid frame size are:
              *
@@ -162,7 +162,7 @@ static int avRecCreateStreamOut(PPDMIHOSTAUDIO pInterface,
 #ifdef VBOX_WITH_LIBOPUS
     PDRVAUDIOVIDEOREC pThis = PDMIHOSTAUDIO_2_DRVAUDIOVIDEOREC(pInterface);
 
-    uint16_t uHz = pStreamOut->Props.uHz;
+    uint32_t uHz = pStreamOut->Props.uHz;
 
     /* Opus only supports certain input sample rates in an efficient manner.
      * So make sure that we use those by resampling the data to the requested rate. */
